WordNet-Affect Lexicon

The WordNet-Affect Lexicon is a hand-curate collection of emotion-related words (nouns, verbs, adjectives, and adverbs), classified as “Positive”, “Negative”, “Neutral”, or “Ambiguous” and categorized into 28 subcategories (“Joy”, “Love”, “Fear”, etc.).

Terms can and do appear in multiple categories.

The original lexicon contains multi-word phrases, but they are excluded here. Also, we removed the term ‘thing’ from the lexicon.

The original WordNet-Affect lexicon is distributed as part of the WordNet Domains project, which is licensed under a Creative Commons Attribution 3.0 Unported License. You are free to share and adapt the lexicon, as long as you give attribution to the original authors.

affect_wordnet

Format

A data frame with one row for each term classification.

Source

https://wndomains.fbk.eu/wnaffect.html

References

Strapparava, C and Valitutti A. (2004). WordNet-Affect: an affective extension of WordNet. Proceedings of the 4th International Conference on Language Resources and Evaluation 1083--1086.